requirements.dev.txt
-r requirements.txt

pytest==8.2.0
pytest-django==4.8.0
black==24.4.2
flake8==7.0.0
